# persist: convert trace batches to use ColumnarRecords