@ruchirK 2021-12-06
Checked offline with @philip-stoev and I'm going to not merge this PR until the currently in-flight release goes out because we're in the following interesting scenario:

* This PR increments `meta` and as a result it is a breaking change to the persistence upgrade tests which need to be repointed to `current`
* However, the persistence upgrade tests are already pointed to `current`, from https://github.com/MaterializeInc/materialize/commit/ea74a4592282a1b947439a24d9beddca6e392825

* Theres currently a release in flight, which doesn't include these changes (as they have not yet been merged), but it does include the commit linked above
* When that release finishes, it will rename the test files to point to the release tag, because the commit linked above will have landed in the release.

This means that merging this PR right now, before the release has completed has the potential to either disturb the release (by somehow messing with the post-release rename) or transiently break tests (because after the tests are renamed they will be broken with respect to this PR). The safest thing to do is just to delay merging this PR until after the release has completed and this seems like a fine thing to do because we don't have any urgent reason to merge this commit, and there's no indication that the release will be further delayed.

If the release gets further delayed, I will rethink what to do. In the meantime I'm marking this PR as a draft to make sure I don't accidentally hit the big green button.

cc: @danhhz