@ruchirK 2022-01-18
hey @danhhz updated the comment and the commit message as requested. I'm not sure what you're thinking of re: persist microbenchmarks because we do have some coverage of this in the encoding benchmarks, which show that encoding trace batches gets ~45% faster with this change. I remember us talking about how the compaction changes aren't covered in the persist microbenchmarks -- could that perhaps be what you're thinking of?

I've spun up a ec2 instance to run the feature benchmarks but I think it takes a while to run everything - if you feel like its no longer needed given the microbenchmark results listed above let me know.