@danhhz 2022-01-18
> I remember us talking about how the compaction changes aren't covered in the persist microbenchmarks -- could that perhaps be what you're thinking of? Edit: oh I also remember us talking about how the indexed write benchmarks never seal - perhaps thats what you're thinking of?

yeah that was it. is the encode_trace benchmark the only one that changed? i'm surprised that end_to_end wasn't also affected