@ruchirK 2022-01-18
`end_to_end` wasn't affected:

```
end_to_end/end_to_end/50MiB
                        time:   [896.01 ms 901.65 ms 907.37 ms]
                        thrpt:  [55.104 MiB/s 55.454 MiB/s 55.803 MiB/s]
                 change:
                        time:   [+1.7575% +3.0083% +4.3387%] (p = 0.00 < 0.05)
                        thrpt:  [-4.1583% -2.9204% -1.7272%]
                        Performance has regressed.
```

I've attached one full run of benchmarks below but `encode_trace` is the only one that is consistently different in the same magnitude / direction compared to main (its 35 - 45% faster than main)
[benchmarks.txt](https://github.com/MaterializeInc/materialize/files/7892987/benchmarks.txt)


eg `encode_batch/unsealed` on my laptop is inherently noisy

and on the same commit can take

```
encode_batch/unsealed/50MiB
                        time:   [54.046 ms 55.609 ms 57.207 ms]
                        thrpt:  [874.03 MiB/s 899.14 MiB/s 925.14 MiB/s]
                 change:
                        time:   [+15.341% +18.950% +22.483%] (p = 0.00 < 0.05)
                        thrpt:  [-18.356% -15.931% -13.301%]
                        Performance has regressed.
```

and 

```
encode_batch/unsealed/50MiB
                        time:   [47.115 ms 47.612 ms 48.118 ms]
                        thrpt:  [1.0148 GiB/s 1.0255 GiB/s 1.0364 GiB/s]
                 change:
                        time:   [-16.900% -14.381% -11.762%] (p = 0.00 < 0.05)
                        thrpt:  [+13.330% +16.796% +20.336%]
                        Performance has improved.
```

(the second set of changes are relative to the "regression" in the first set of changes)