@ruchirK 2022-01-19
feature benchmark results

```
feature-benchmark_mzdata
NAME                      |    THIS     |    OTHER    |  Regression?  | 'THIS' is:
----------------------------------------------------------------------------------------------------
Sleep                     |       0.102 |       0.102 |      no       | 0.0 pct   faster
FastPathFilterNoIndex     |       0.533 |       0.562 |      no       | 5.1 pct   faster
FastPathFilterIndex       |       0.167 |       0.169 |      no       | 1.0 pct   faster
FastPathOrderByLimit      |       0.280 |       0.283 |      no       | 1.1 pct   faster
Insert                    |       2.820 |       2.607 |      no       | 8.1 pct   slower
Update                    |       2.117 |       1.784 |    !!YES!!    | 18.6 pct   slower
InsertAndSelect           |       3.613 |       3.343 |      no       | 8.1 pct   slower
OrderBy                   |       7.105 |       6.851 |      no       | 3.7 pct   slower
CountDistinct             |       1.515 |       1.525 |      no       | 0.7 pct   faster
MinMax                    |       2.817 |       2.784 |      no       | 1.2 pct   slower
GroupBy                   |       8.508 |       8.308 |      no       | 2.4 pct   slower
CrossJoin                 |       2.036 |       2.076 |      no       | 1.9 pct   faster
Retraction                |       2.254 |       2.194 |      no       | 2.7 pct   slower
CreateIndex               |       0.206 |       0.199 |      no       | 3.9 pct   slower
DeltaJoin                 |       0.354 |       0.352 |      no       | 0.8 pct   slower
DifferentialJoin          |       0.687 |       0.673 |      no       | 2.1 pct   slower
FinishOrderByLimit        |       0.145 |       0.151 |      no       | 4.2 pct   faster
KafkaRaw                  |       1.170 |       1.178 |      no       | 0.7 pct   faster
KafkaUpsert               |       2.041 |       1.926 |      no       | 6.0 pct   slower
KafkaUpsertUnique         |       2.001 |       2.149 |      no       | 6.9 pct   faster
KafkaRecovery             |      21.035 |      17.617 |    !!YES!!    | 19.4 pct   slower
ExactlyOnce               |       4.824 |       4.536 |      no       | 6.3 pct   slower
```

investigating `Update` and `KafkaRecovery` in more detail