@danhhz 2022-01-19
> investigating Update and KafkaRecovery in more detail

what are the raw data points?