@ruchirK 2022-01-25
> Can you think of anything in this PR that would have changed encode_batch/unsealed (unsealed not trace) or is that improvement just noise?

So, to test this out I ran a couple of iterations of just that benchmark, and nothing else with:

```
cargo bench --bench writer -- encode_batch/unsealed
```

(I had a couple of docker containers for kafka/schema registry up, but nothing else was up and a skim over `htop` indicated that the system was unloaded for memory and ram).

on my branch the output looked like:
```
ubuntu@ip-10-1-25-140:~/materialize/src/persist$ cargo bench --bench writer -- encode_batch/unsealed
    Finished bench [optimized + debuginfo] target(s) in 0.26s
     Running unittests (/home/ubuntu/materialize/target/release/deps/writer-f134d4e0ee414587)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
encode_batch/unsealed/50MiB
                        time:   [40.475 ms 40.502 ms 40.529 ms]
                        thrpt:  [1.2048 GiB/s 1.2056 GiB/s 1.2064 GiB/s]
                 change:
                        time:   [+0.1519% +0.2537% +0.3561%] (p = 0.00 < 0.05)
                        thrpt:  [-0.3548% -0.2531% -0.1516%]
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild

MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
ubuntu@ip-10-1-25-140:~/materialize/src/persist$ cargo bench --bench writer -- encode_batch/unsealed
    Finished bench [optimized + debuginfo] target(s) in 0.26s
     Running unittests (/home/ubuntu/materialize/target/release/deps/writer-f134d4e0ee414587)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
encode_batch/unsealed/50MiB
                        time:   [40.313 ms 40.344 ms 40.375 ms]
                        thrpt:  [1.2094 GiB/s 1.2103 GiB/s 1.2112 GiB/s]
                 change:
                        time:   [-0.4925% -0.3883% -0.2889%] (p = 0.00 < 0.05)
                        thrpt:  [+0.2898% +0.3898% +0.4949%]
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) low mild

MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
ubuntu@ip-10-1-25-140:~/materialize/src/persist$ cargo bench --bench writer -- encode_batch/unsealed
    Finished bench [optimized + debuginfo] target(s) in 0.26s
     Running unittests (/home/ubuntu/materialize/target/release/deps/writer-f134d4e0ee414587)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
encode_batch/unsealed/50MiB
                        time:   [40.523 ms 40.554 ms 40.584 ms]
                        thrpt:  [1.2031 GiB/s 1.2040 GiB/s 1.2050 GiB/s]
                 change:
                        time:   [+0.4103% +0.5196% +0.6270%] (p = 0.00 < 0.05)
                        thrpt:  [-0.6231% -0.5170% -0.4086%]
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) low mild
```


The first run is a regression against the 23 ms reported in the previous run, but the results are pretty stable at around 40ms. On main, this looks like

```
ubuntu@ip-10-1-25-140:~/materialize/src/persist$ git checkout ruchir/main
Previous HEAD position was 8b85b9cfa persist: convert trace batches to use ColumnarRecords
HEAD is now at 101f81af5 Merge pull request #9988 from mjibson/tablefuncs
ubuntu@ip-10-1-25-140:~/materialize/src/persist$ cargo bench --bench writer -- encode_batch/unsealed
   Compiling persist v0.0.0 (/home/ubuntu/materialize/src/persist)
    Finished bench [optimized + debuginfo] target(s) in 31.62s
     Running unittests (/home/ubuntu/materialize/target/release/deps/writer-f134d4e0ee414587)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
encode_batch/unsealed/50MiB
                        time:   [40.564 ms 40.591 ms 40.619 ms]
                        thrpt:  [1.2021 GiB/s 1.2029 GiB/s 1.2037 GiB/s]
                 change:
                        time:   [-0.0147% +0.0912% +0.1952%] (p = 0.08 > 0.05)
                        thrpt:  [-0.1948% -0.0911% +0.0147%]
                        No change in performance detected.

MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
ubuntu@ip-10-1-25-140:~/materialize/src/persist$ cargo bench --bench writer -- encode_batch/unsealed
    Finished bench [optimized + debuginfo] target(s) in 0.26s
     Running unittests (/home/ubuntu/materialize/target/release/deps/writer-f134d4e0ee414587)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
encode_batch/unsealed/50MiB
                        time:   [40.284 ms 40.318 ms 40.353 ms]
                        thrpt:  [1.2100 GiB/s 1.2111 GiB/s 1.2121 GiB/s]
                 change:
                        time:   [-0.7761% -0.6718% -0.5682%] (p = 0.00 < 0.05)
                        thrpt:  [+0.5715% +0.6764% +0.7822%]
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
ubuntu@ip-10-1-25-140:~/materialize/src/persist$ cargo bench --bench writer -- encode_batch/unsealed
    Finished bench [optimized + debuginfo] target(s) in 0.26s
     Running unittests (/home/ubuntu/materialize/target/release/deps/writer-f134d4e0ee414587)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
encode_batch/unsealed/50MiB
                        time:   [40.575 ms 40.602 ms 40.630 ms]
                        thrpt:  [1.2018 GiB/s 1.2026 GiB/s 1.2034 GiB/s]
                 change:
                        time:   [+0.5964% +0.7047% +0.8177%] (p = 0.00 < 0.05)
                        thrpt:  [-0.8111% -0.6998% -0.5928%]
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild

MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
MZ_PERSIST_RECORD_COUNT=819200 MZ_PERSIST_RECORD_SIZE_BYTES=64 MZ_PERSIST_BATCH_MAX_COUNT=131072
```

which is basically the same. Its extremely suspicious that running only this benchmark causes a slower result than running this benchmark with all of the other benchmarks (each benchmark run is 200 iterations which feels like enough time for caches to stabilize but perhaps not?). Either way, the improvement seems like noise?